add real_gaussian_integral function (#371)
**Context:** Add the code for real_gaussian_integral to help the
from_quadrature() function.

**Description of the Change:** The mathematical formula corresponds to
the one on the overleaf Appendix.

TODO list:
- [x]  Add a test for this? (No idea how to test this.)
- [x]  Move the complex gaussian integral into this file as well?

**Benefits:** We can deal with the basis change with a real integral.

**Possible Drawbacks:** We may need a for loop outside to loop on the
integral with different values of b. (Need to be discussed!)

**Related GitHub Issues:**

---------
