+++ doc/code/physics/utils/gaussian_integral.rst
@@ -1,0 +1,8 @@
+Calculations on Gaussian Integrals
+================================
+
+.. currentmodule:: mrmustard.physics.bargmann
+
+.. automodapi:: mrmustard.physics.bargmann
+    :no-heading:
+    :include-all-objects:
