+++ mrmustard/physics/representations.py
@@ -26,7 +26,11 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from mrmustard import math, settings
-from mrmustard.physics import bargmann
+from mrmustard.physics.gaussian_integrals import (
+    contract_two_Abc,
+    reorder_abc,
+    complex_gaussian_integral,
+)
 from mrmustard.physics.ansatze import Ansatz, PolyExpAnsatz, ArrayAnsatz
 from mrmustard.utils.typing import (
     Batch,
@@ -289,7 +293,7 @@ class Bargmann(Representation):
             )
         A, b, c = [], [], []
         for Abci in zip(self.A, self.b, self.c):
-            Aij, bij, cij = bargmann.complex_gaussian_integral(Abci, idx_z, idx_zconj, measure=-1.0)
+            Aij, bij, cij = complex_gaussian_integral(Abci, idx_z, idx_zconj, measure=-1.0)
             A.append(Aij)
             b.append(bij)
             c.append(cij)
@@ -316,7 +320,7 @@ class Bargmann(Representation):
         Returns:
             The reordered Bargmann object.
         """
-        A, b, c = bargmann.reorder_abc((self.A, self.b, self.c), order)
+        A, b, c = reorder_abc((self.A, self.b, self.c), order)
         return self.__class__(A, b, c)
 
     def plot(
@@ -440,14 +444,7 @@ class Bargmann(Representation):
         Abc = []
         for A1, b1, c1 in zip(self.A, self.b, self.c):
             for A2, b2, c2 in zip(other.A, other.b, other.c):
-                Abc.append(
-                    bargmann.contract_two_Abc(
-                        (A1, b1, c1),
-                        (A2, b2, c2),
-                        idx_s,
-                        idx_o,
-                    )
-                )
+                Abc.append(contract_two_Abc((A1, b1, c1), (A2, b2, c2), idx_s, idx_o))
 
         A, b, c = zip(*Abc)
         return Bargmann(A, b, c)
