+++ tests/test_physics/test_representations.py
@@ -20,7 +20,7 @@ import pytest
 from mrmustard import math, settings
 from mrmustard.physics.converters import to_fock
 from mrmustard.physics.triples import displacement_gate_Abc, attenuator_Abc
-from mrmustard.physics.bargmann import contract_two_Abc, complex_gaussian_integral
+from mrmustard.physics.gaussian_integrals import contract_two_Abc, complex_gaussian_integral
 from mrmustard.physics.representations import Bargmann, Fock
 from ..random import Abc_triple
